관리자 서비스 카운트 수정

+++ src/main/java/com/bugjeogbugjeog/app/bugjeogbugjeog/service/NoticeService.java
@@ -30,5 +30,5 @@ public class NoticeService {
     }
 
     /* 카운트 */
-    public int count(){return noticeDAO.count();}
+    public Long count(){return noticeDAO.count();}
 }
